NEW FEATURE! Accessing map data efficiently (see details) + Other features & improvements
On all previous releases of PyOpenNI & ONIPY, each time the user requested the
depth map of a DepthGenerator, we *copied each pixel* of the map into a Python
tuple, and returned that. This was very slow and the user had to manually
calculate the index of the pixel he wanted (x + y*width).

Now, DepthGenerator(s) have the property ‘map’, which returns a DepthMap
wrapping a *reference* (pointer) to the data. It's very easy to use:

    my_map[x,y]

It will calculate the index for you, check the array bounds to ensure this pixel
actually exists, and return it "on demand".

I've made it flexible so it can also be used as a plain array:

    my_map[i]

This allows easy migration of older code.
Switch to it now, you'll gain a LOT of speed!

BONUS: The DepthMap object also contains the dimensions of the map:

    my_map.width
    my_map.height
    my_map.size  # (width, height) tuple

    len(my_map) # Length of the array, equivalent to width*height

See the help(...) on DepthMap for more info.